Move funcs with static filters for pylint error C0302: Too many lines in module
